Move replication lag to common
